update users on frontend, corrections on backend

+++ backend/src/routes/user_routes.rs
@@ -5,14 +5,21 @@ use crate::models::jwt_model::JWTInternal;
 use crate::models::user_model::{NewUser, User};
 use crate::StorageState;
 use actix_web::web;
+use log::error;
 use serde::{Deserialize, Serialize};
+use crate::models::group_user_model::GroupUser;
+use crate::models::role_model::Role;
+use crate::models::role_user_model::RoleUser;
 
 pub(crate) async fn create_user(
     form_data: web::Json<NewUser>,
     db: web::Data<StorageState>,
 ) -> Result<web::Json<User>, ApiError> {
     let mut db = try_get_connection(&db)?;
-    Ok(web::Json(User::create(&mut db, &form_data.0)?))
+    let user = User::create(&mut db, &form_data.0)?;
+    let public_group = &Group::get(&mut db, 1)?;
+    GroupUser::add_user_to_group(&mut db, &user, public_group)?;  // group public with id 0 should always exist
+    Ok(web::Json(user))
 }
 
 pub(crate) async fn all_users(
@@ -50,7 +57,16 @@ pub(crate) async fn update_user(
         user_retrieved.login = new_login;
     };
     if let Some(new_hash) = user_update_payload.new_hash.clone() {
-        user_retrieved.hash = new_hash;
+        if new_hash.len() < 4 {
+            return Err(ApiError::User);
+        }
+        user_retrieved.hash = match bcrypt::hash(new_hash, 12) {
+            Ok(new_hash) => new_hash,
+            Err(e) => {
+                error!("{e:?}");
+                return Err(ApiError::Internal);
+            },
+        };
     };
     User::update_user(&mut db, &user_retrieved)?;
     Ok("updated.")
@@ -63,9 +79,14 @@ pub(crate) async fn delete_user(
     let mut db = try_get_connection(&db)?;
     let uid = path.into_inner();
     let user = User::get(&mut db, uid)?;
-    JWTInternal::invalidate_user(&mut db, &user)?;
-    User::delete_user(&mut db, &user)?;
-    Ok("deleted.")
+
+    if RoleUser::roles_from_user(&mut db, &user)? == Role::from("root")? {
+        Err(ApiError::CantDeleteRoot)
+    } else {
+        JWTInternal::invalidate_user(&mut db, &user)?;
+        User::delete_user(&mut db, &user)?;
+        Ok("deleted.")
+    }
 }
 
 pub(crate) async fn get_user_groups(
